+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   `@mdi/font` update to the version 5.4.55
+-   `@mdi/js` update to the version 5.4.55
+
 ## [0.25.6][] - 2020-07-31
 
 ### Fixed